Post Section 4 Lectures - Improving Cheese 3D Shooter - Day 5
Menus have been fully added, bugs have been fixed and now it needs to be built for the web. I have built a PC build but Unity keeps failing to build for the web and the pc build.
I have a solution now thanks to a person the discord but I want to make this commit just in case moving these folders goes badly.